requirements.txt
streamlit
streamlit-lottie
requests
pillow